scroller in the pipeline

+++ src/constants.h
@@ -11,7 +11,7 @@ extern const int TPH_TICKS_PER_CLOCK; // PPQN = 192
 extern const char *const ASSETS_DIRECTORY;
 extern const char *const USER_DIRECTORY;
 
-constexpr int TPH_RACK_COUNT = 4; // keep small for easy debugging..
+constexpr uint8_t TPH_RACK_COUNT = 4; // keep small for easy debugging..
 constexpr int TPH_RACK_RENDER_SIZE = 64;
 constexpr int TPH_RACK_BUFFER_SIZE = 128; // x2 since stereo
 constexpr int LUT_SIZE = 16384;
